move the carsListRequest function to logic/carsRequests.js
